Support configuring avro compression in ManifestLists
The iceberg table properties are never actually propagated to Avro
writers in Iceberg. As a result of this, the configuration for Avro
compression can never actually be altered, and all avro files are
compressed with gzip.

In this change, let's fix support adding table properties to
ManifestList avro files.

Change-Id: I6596db6724dea13f5979c9380232f23dcdc3cfef
Reviewed-on: https://gerrit.trading.imc.intra/c/data-engineering/iceberg/+/660815
Tested-by: Teamcity
Reviewed-by: Shiva Ganapathy <shiva.ganapathy@imc.com>
Reviewed-by: Lorenzo Siega Battel <lorenzo.siegabattel@imc.com>
Reviewed-by: Alexander Borodin <alex.borodin@imc.com>
Static-Analysis: Teamcity